requirements.txt
streamlit>=1.28.0
psycopg2-binary
pandas
python-dotenv
langchain
langchain-google-genai
chromadb
google-generativeai>=0.3.0
tqdm   
langchain-core  
langchain-community
sentence-transformers>=2.2.2
plotly
streamlit-lottie
fastapi
uvicorn
psycopg2-binary
pandas
google-generativeai
langchain-community
python-dotenv
reportlab
matplotlib